+++ sql/NexoraDB/Tables/dbo.ActiveSessions.sql
@@ -4,6 +4,8 @@ ALTER TABLE [dbo].[ActiveSessions] DROP CONSTRAINT [DF_ActiveSessions_CreatedAt]
 GO
 DROP INDEX [IX_ActiveSessions_UserID] ON [dbo].[ActiveSessions]
 GO
+DROP INDEX [IX_ActiveSessions_LastSeenAt] ON [dbo].[ActiveSessions]
+GO
 DROP TABLE [dbo].[ActiveSessions]
 GO
 SET ANSI_NULLS ON
@@ -15,11 +17,17 @@ CREATE TABLE [dbo].[ActiveSessions](
 	[UserID] [int] NOT NULL,
 	[CreatedAt] [datetime] NOT NULL,
 	[IPAddress] [nvarchar](45) NULL,
+	[LastSeenAt] [datetime] NOT NULL,
 PRIMARY KEY CLUSTERED 
 (
 	[SessionID] ASC
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
 ) ON [PRIMARY]
+GO
+CREATE NONCLUSTERED INDEX [IX_ActiveSessions_LastSeenAt] ON [dbo].[ActiveSessions]
+(
+	[LastSeenAt] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, SORT_IN_TEMPDB = OFF, DROP_EXISTING = OFF, ONLINE = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
 GO
 CREATE NONCLUSTERED INDEX [IX_ActiveSessions_UserID] ON [dbo].[ActiveSessions]
 (